Split tests for request headers
These tests all inspect headers in a request received from edge by origin.

It leaves only a handful of tests left in the original uncategorised
"cdn_acceptance_test.go" file.
